extract hover selector to its own class
