+++ scopes/react/mouse-hover-selector/index.ts
