+++ scopes/react/mouse-hover-selector/mouse-hover-selector.tsx
